// iot-service/iot-angular/src/sdk/core/models/forwarding-database-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ForwardingDatabaseResponseDto {

  /**
   * Batch size
   */
  batchSize: number;

  /**
   * Created timestamp
   */
  createdAt: string;

  /**
   * Database name
   */
  databaseName: string;

  /**
   * Database type
   */
  dbType: string;

  /**
   * Is database config enabled
   */
  enabled: boolean;

  /**
   * Database host
   */
  host: string;

  /**
   * Owner ID
   */
  idOwner: string;

  /**
   * Database config unique identifier
   */
  idOwnerForwardingDb: string;

  /**
   * Database config label
   */
  label: string;

  /**
   * Last delivery timestamp
   */
  lastDeliveryAt?: string;

  /**
   * Last error message
   */
  lastError?: string;

  /**
   * Last delivery status
   */
  lastStatus?: string;

  /**
   * Password (masked)
   */
  passwordCipher: string;

  /**
   * Database port
   */
  port: number;

  /**
   * Target schema
   */
  targetSchema?: string;

  /**
   * Target table name
   */
  targetTable: string;

  /**
   * Updated timestamp
   */
  updatedAt: string;

  /**
   * Database username
   */
  username: string;

  /**
   * Write mode
   */
  writeMode: string;
}